+++ src/render/engine.rs
@@ -221,11 +221,9 @@ impl Engine {
             self.widget_cache.tick(self.layout_cache.get_layout_cache());
             self.layout_cache
                 .do_layout(self.widget_cache.borrow_cache());
+            self.widget_cache.draw_loop(&mut canvas);
 
-            // Flip screen display canvas only if the draw loop requires a refresh.
-            if self.widget_cache.draw_loop(&mut canvas) {
-                canvas.present();
-            }
+            canvas.present();
 
             // This obeys thread sleep time.
             let now = SystemTime::now()